requirements.txt
flask
google-generativeai
google-cloud-bigquery
google-cloud-logging
pandas
db_dtypes
